+++ src/materials/glass.hpp
@@ -2,7 +2,7 @@
 
 class Glass: public Material {
 public:
-    double n_i;
+    float n_i;
     Color c;
     Color getColor(vec4 & pos);
     ray scatter(ray & inRay, vec4 & pos, vec4 & normal);
